refactor useToken file to live in useToken.ts
